refactor: modularize dropdown and signup logic, add cache manager, complete JSDoc coverage
- extract cacheManager from overview (overview.js 397 LOC)
- split dropdownManager into contactDropdownManager and categoryDropdownManager with re-exports
- split signup validation into dedicated signupValidation module
- add full JSDoc to credentials (not pushed), overview, errorHandler, login, database
- add @fileoverview and complete documentation to all new modules
- ensure all files <= 400 LOC
- achieve 100% JSDoc coverage (0 undocumented functions)
