+++ js/dropdownManager.js
@@ -1,199 +1,40 @@
-import {generateContactOptionHTML, getContactChipHTML, getCategoryDropdownOptionsHTML} from "./template.js";
-import {database} from "./database.js";
-import {ref, onValue} from "https://www.gstatic.com/firebasejs/12.4.0/firebase-database.js";
+/**
+ * @fileoverview Core dropdown manager handling generic dropdown UI logic (toggle, close, clear),
+ * event delegation for both contact and category dropdowns, and initialization.
+ * Re-exports all public symbols from contactDropdownManager and categoryDropdownManager
+ * so that existing import paths remain unchanged.
+ */
 
-let contacts = [];
-let selectedContacts = [];
-let selectedCategory = null;
+import {
+    loadContacts,
+    populateContactsDropdown,
+    selectContact as _selectContact,
+    filterOptions as _filterOptions,
+    setActiveContainerRef
+} from "./contactDropdownManager.js";
+
+import {
+    populateCategoriesDropdown,
+    selectCategory as _selectCategory,
+    setSelectedCategory as _setSelectedCategory
+} from "./categoryDropdownManager.js";
+
+import {setSelectedContacts as _setSelectedContacts} from "./contactDropdownManager.js";
+
+/** @type {HTMLElement|null} The currently active form container */
 let activeContainer = null;
-let contactsLoaded = false;
-let contactsLoadedPromise = null;
+
+/**
+ * WeakMap storing AbortControllers keyed by container elements.
+ * Used to clean up event listeners when a container is re-initialized.
+ * @type {WeakMap<HTMLElement, AbortController>}
+ */
 const dropdownAbortControllers = new WeakMap();
 
-const categories = [
-    {id: 'technical', name: 'Technical Task'},
-    {id: 'user-story', name: 'User Story'}
-];
-
 /**
- * Load contacts from database for dropdown (with caching)
- * Firebase listener is registered only once, subsequent calls use cached data
- * @returns {Promise<void>} - Promise that resolves when contacts are loaded
- */
-function loadContacts() {
-    if (contactsLoaded && contacts.length > 0) return Promise.resolve();
-
-    if (contactsLoadedPromise) return contactsLoadedPromise;
-
-    contactsLoadedPromise = new Promise((resolve) => {
-        const contactsRef = ref(database, 'contacts');
-
-        onValue(contactsRef, (snapshot) => {
-            if (!snapshot.exists()) return;
-            contacts = Object.values(snapshot.val());
-            contactsLoaded = true;
-            if (!activeContainer) return;
-            populateContactsDropdown(activeContainer);
-            resolve();
-        });
-    });
-    return contactsLoadedPromise;
-}
-
-/**
- * Populate the contacts dropdown with available contacts
- * @param {HTMLElement} container - The container element to scope queries
- */
-function populateContactsDropdown(container) {
-    const dropdownContent = container.querySelector('.contact-dropdown-content');
-    if (!dropdownContent) return;
-    dropdownContent.innerHTML = '';
-    const sortedContacts = contacts.sort((a, b) => a.name.localeCompare(b.name));
-    sortedContacts.forEach(contact => {
-        const contactOption = document.createElement('div');
-        contactOption.className = 'contact-option';
-        contactOption.setAttribute('data-contact-id', contact.id);
-        contactOption.innerHTML = generateContactOptionHTML(contact);
-        dropdownContent.appendChild(contactOption);
-    });
-}
-
-/**
- * Select/deselect a contact
- * @param {string} contactId - The contact ID
- * @param {HTMLElement} container - The container element to scope queries
- */
-export function selectContact(contactId, container) {
-    const isSelected = selectedContacts.some(c => c.id === contactId);
-    toggleContactSelection(contactId, isSelected, container);
-    updateSelectedContactsDisplay(container);
-}
-
-/**
- * Toggles contact selection state
- * @param {string} contactId - The contact ID
- * @param {object|null} isSelected - The currently selected contact or null
- * @param {HTMLElement} container - The container element to scope queries
- * @returns {void}
- */
-function toggleContactSelection(contactId, isSelected, container) {
-    const {contact, contactOption, checkbox} = getContactElements(contactId, container);
-    if (!contact || !contactOption || !checkbox) return;
-
-    if (isSelected) {
-        selectedContacts = selectedContacts.filter(contact => contact.id !== contactId);
-        contactOption.classList.remove('selected');
-        checkbox.checked = false;
-    } else {
-        selectedContacts.push(contact);
-        contactOption.classList.add('selected');
-        checkbox.checked = true;
-    }
-}
-
-/**
- * Get contact elements by ID
- * @param {string} contactId - The contact ID
- * @param {HTMLElement} container - The container element to scope queries
- * @returns {object} - Object containing contact, option, and checkbox elements
- */
-function getContactElements(contactId, container) {
-    const contact = contacts.find(c => c.id === contactId);
-    const contactOption = container.querySelector(`[data-contact-id="${contactId}"]`);
-    const checkbox = contactOption?.querySelector(`input[type="checkbox"]`);
-
-    return {contact, contactOption, checkbox};
-}
-
-/**
- * Update the display of selected contacts
- * @param {HTMLElement} container - The container element to scope queries
- */
-function updateSelectedContactsDisplay(container) {
-    const dropzone = container.querySelector('.dropzone');
-    if (!dropzone) return;
-
-    dropzone.innerHTML = '';
-
-    if (selectedContacts.length >= 5) {
-        for (let i = 0; i < selectedContacts.length; i++) {
-            if (i < 5) appendContactChip(selectedContacts[i], dropzone);
-        }
-
-        const overflowCount = selectedContacts.length - 5;
-        if (overflowCount > 0) generateOverflowChip(overflowCount, dropzone);
-    } else {
-        selectedContacts.forEach(contact => {
-            appendContactChip(contact, dropzone);
-        });
-    }
-}
-
-/**
- * Generates an overflow chip showing the count of additional contacts
- * @param {number} count - The number of additional contacts to display
- * @param {HTMLElement} dropzone - The dropzone element to append the chip to
- * @returns {void}
- */
-function generateOverflowChip(count, dropzone) {
-    const overflowChip = document.createElement('div');
-    overflowChip.className = 'contact-chip overflow';
-    overflowChip.textContent = `+${count}`;
-    dropzone.appendChild(overflowChip);
-}
-
-/**
- * Appends a contact chip element to the dropzone
- * @param {Object} contact - The contact object containing contact information
- * @param {HTMLElement} dropzone - The dropzone element to append the chip to
- * @returns {void}
- */
-function appendContactChip(contact, dropzone) {
-    const contactChip = document.createElement('div');
-    contactChip.className = 'contact-chip';
-    contactChip.innerHTML = getContactChipHTML(contact);
-    dropzone.appendChild(contactChip);
-}
-
-/**
- * Populate the category dropdown with available categories
- * @param {HTMLElement} container - The container element to scope queries
- */
-function populateCategoriesDropdown(container) {
-    const dropdownContent = container.querySelector('.category-dropdown-content');
-    if (!dropdownContent) return;
-
-    dropdownContent.innerHTML = getCategoryDropdownOptionsHTML();
-}
-
-/**
- * Select a category
- * @param {string} categoryId - The category ID
- * @param {HTMLElement} container - The container element to scope queries
- */
-export function selectCategory(categoryId, container) {
-    const category = categories.find(c => c.id === categoryId);
-    const categoryDisplay = container.querySelector('.category-display');
-    if (!category) return;
-
-    selectedCategory = category;
-    categoryDisplay.textContent = category.name;
-
-    container.querySelectorAll('.category-option').forEach(option => {
-        option.classList.remove('selected');
-    });
-
-    const selectedOption = container.querySelector(`[data-category-id="${categoryId}"]`);
-    if (selectedOption) {
-        selectedOption.classList.add('selected');
-    }
-    toggleDropdown('category', false, container);
-}
-
-/**
- * Toggle dropdown visibility
+ * Toggles a dropdown's visibility. Opens if closed (or forced), closes if open.
  * @param {string} type - Type of dropdown ('contact' or 'category')
- * @param {boolean} forceOpen - Force dropdown to open
+ * @param {boolean} [forceOpen=false] - Force the dropdown to open
  * @param {HTMLElement} container - The container element to scope queries
  * @returns {void}
  */
@@ -205,20 +46,17 @@ export function toggleDropdown(type, forceOpen = false, container) {
         closeAllDropdowns(type, container);
         dropdownHeader.classList.add('active');
         dropdownContent.classList.add('active');
-        if (type === 'contact') {
-            const searchInput = container.querySelector('.contact-search-input');
-            if (searchInput) searchInput.focus();
-        }
+        if (type === 'contact') container.querySelector('.contact-search-input').focus();
     } else {
         clearDropdown(type, container);
     }
 }
 
 /**
- * Get dropdown elements by type
- * @param {string} type - Type of dropdown
+ * Retrieves the wrapper, content, and header elements for a dropdown type.
+ * @param {string} type - Type of dropdown ('contact' or 'category')
  * @param {HTMLElement} container - The container element to scope queries
- * @returns {object} - Object containing wrapper, content, and header elements
+ * @returns {{wrapper: HTMLElement|null, dropdownContent: HTMLElement|null, dropdownHeader: HTMLElement|null}}
  */
 function getDropdownElements(type, container) {
     const wrapper = container.querySelector(`.${type}-dropdown-wrapper`);
@@ -229,8 +67,9 @@ function getDropdownElements(type, container) {
 }
 
 /**
- * Close all dropdowns except the specified one
- * @param {string|null} except - Type to exclude from closing
+ * Closes all dropdowns except the specified type.
+ * Resets the contact search filter when closing the contact dropdown.
+ * @param {string|null} [except=null] - Dropdown type to exclude from closing
  * @param {HTMLElement} container - The container element to scope queries
  * @returns {void}
  */
@@ -238,52 +77,30 @@ function closeAllDropdowns(except = null, container) {
     const types = ['contact', 'category'];
 
     types.forEach(type => {
-        if (type !== except) {
-            clearDropdown(type, container);
-            if (type === 'contact') filterOptions('contact', container);
-        }
+        if (type === except) return;
+
+        clearDropdown(type, container);
+        if (type === 'contact') _filterOptions(container, toggleDropdown);
     });
 }
 
 /**
- * Clear the dropdown section
- * @param {string} type - Type of dropdown to clear
+ * Clears (closes) a single dropdown section and resets the contact search input.
+ * @param {string} type - Type of dropdown to clear ('contact' or 'category')
  * @param {HTMLElement} container - The container element to scope queries
  * @returns {void}
  */
 function clearDropdown(type, container) {
     const {dropdownContent, dropdownHeader} = getDropdownElements(type, container);
     if (!dropdownContent || !dropdownHeader) return;
-    const searchInput = container.querySelector('.contact-search-input');
+    container.querySelector('.contact-search-input').innerText = '';
 
     dropdownContent.classList.remove('active');
     dropdownHeader.classList.remove('active');
-    if (searchInput) searchInput.value = '';
 }
 
 /**
- * Filter options based on search input
- * @param {string} type - Type of dropdown to filter
- * @param {HTMLElement} container - The container element to scope queries
- * @returns {void}
- */
-export function filterOptions(type, container) {
-    if (type === 'contact') {
-        const searchInput = container.querySelector('.contact-search-input');
-        const searchTerm = searchInput ? searchInput.value.toLowerCase() : '';
-        const contactOptions = container.querySelectorAll('.contact-option');
-
-        contactOptions.forEach(option => {
-            const name = option.querySelector('.contact-option-name').textContent.toLowerCase();
-            name.includes(searchTerm) ? option.style.display = 'flex' : option.style.display = 'none';
-        });
-
-        if (searchTerm.length > 0) toggleDropdown('contact', true, container);
-    }
-}
-
-/**
- * Close dropdown when clicking outside
+ * Closes any open dropdowns when a click occurs outside of them.
  * @param {MouseEvent} event - The mouse event
  * @param {HTMLElement} container - The container element to scope queries
  * @returns {void}
@@ -297,17 +114,18 @@ function closeDropdownOnClickOutside(event, container) {
 }
 
 /**
- * Setup event delegation for dropdowns
+ * Sets up click and input event delegation for all dropdown interactions.
  * @param {HTMLElement} container - The container element to scope events
- * @param {AbortSignal} signal - AbortSignal for cleanup
+ * @param {AbortSignal|null} [signal=null] - AbortSignal for cleanup
+ * @returns {void}
  */
 function setupDropdownEventDelegation(container, signal = null) {
     const options = signal ? {signal} : {};
 
     container.addEventListener('click', (event) => {
-        const {contactHeader, searchInput, categoryHeader, contactOption, categoryOption} = closestEvents(event)
+        const {contactHeader, searchInput, categoryHeader, contactOption, categoryOption} = closestEvents(event);
 
-        setupToggleDropdownEventDelegation(event, contactHeader, searchInput, container)
+        setupToggleDropdownEventDelegation(event, contactHeader, searchInput, container);
 
         if (categoryHeader) return toggleDropdown('category', false, container);
         setupContactOptionDelegation(contactOption, container, event);
@@ -317,9 +135,9 @@ function setupDropdownEventDelegation(container, signal = null) {
 }
 
 /**
- * Gets the closest matching elements for dropdown event handling
+ * Gets the closest matching elements for dropdown event handling.
  * @param {MouseEvent} event - The mouse event to analyze
- * @returns {Object} Object containing the closest header, input, and option elements
+ * @returns {{contactHeader: HTMLElement|null, searchInput: HTMLElement|null, categoryHeader: HTMLElement|null, contactOption: HTMLElement|null, categoryOption: HTMLElement|null}}
  */
 function closestEvents(event) {
     const contactHeader = event.target.closest('.contact-dropdown-header');
@@ -331,7 +149,7 @@ function closestEvents(event) {
 }
 
 /**
- * Handles contact option click delegation
+ * Handles contact option click delegation.
  * @param {HTMLElement|null} contactOption - The clicked contact option element
  * @param {HTMLElement} container - The container element to scope queries
  * @param {MouseEvent} event - The mouse event
@@ -342,11 +160,11 @@ function setupContactOptionDelegation(contactOption, container, event) {
     if (isClickOnCheckboxArea(event)) return;
 
     const contactId = contactOption.getAttribute('data-contact-id');
-    if (contactId) selectContact(contactId, container);
+    if (contactId) _selectContact(contactId, container);
 }
 
 /**
- * Checks if the click event occurred on the checkbox area
+ * Checks if the click event occurred on the checkbox area.
  * @param {MouseEvent} event - The mouse event to check
  * @returns {boolean} True if click was on checkbox area, false otherwise
  */
@@ -355,7 +173,7 @@ function isClickOnCheckboxArea(event) {
 }
 
 /**
- * Sets up toggle dropdown event delegation for contact header clicks
+ * Sets up toggle dropdown event delegation for contact header clicks.
  * @param {MouseEvent} event - The mouse event
  * @param {HTMLElement|null} contactHeader - The contact header element
  * @param {HTMLElement|null} searchInput - The search input element
@@ -374,7 +192,7 @@ function setupToggleDropdownEventDelegation(event, contactHeader, searchInput, c
 }
 
 /**
- * Handles category option click delegation
+ * Handles category option click delegation.
  * @param {HTMLElement|null} categoryOption - The clicked category option element
  * @param {HTMLElement} container - The container element to scope queries
  * @returns {void}
@@ -383,13 +201,13 @@ function setupCategoryOptionDelegation(categoryOption, container) {
     if (categoryOption) {
         const categoryId = categoryOption.getAttribute('data-category-id');
         if (categoryId) {
-            selectCategory(categoryId, container);
+            _selectCategory(categoryId, container);
         }
     }
 }
 
 /**
- * Sets up event delegation for contact-related events (search and checkbox)
+ * Sets up event delegation for contact-related events (search input and checkbox changes).
  * @param {HTMLElement} container - The container element to scope events
  * @param {Object} options - Event listener options including AbortSignal
  * @returns {void}
@@ -400,7 +218,7 @@ function setupContactEventDelegation(container, options) {
 }
 
 /**
- * Handles search input events for filtering contacts
+ * Handles search input events for filtering contacts.
  * @param {HTMLElement} container - The container element to scope events
  * @param {Object} options - Event listener options including AbortSignal
  * @returns {void}
@@ -408,12 +226,12 @@ function setupContactEventDelegation(container, options) {
 function handleSearchInput(container, options) {
     container.addEventListener('input', (event) => {
         const isSearchInput = event.target.closest('.contact-search-input');
-        if (isSearchInput) filterOptions('contact', container);
+        if (isSearchInput) _filterOptions(container, toggleDropdown);
     }, options);
 }
 
 /**
- * Handles checkbox change events for contact selection
+ * Handles checkbox change events for contact selection.
  * @param {HTMLElement} container - The container element to scope events
  * @param {Object} options - Event listener options including AbortSignal
  * @returns {void}
@@ -426,12 +244,13 @@ function handleCheckboxChange(container, options) {
 
         const contactOption = checkbox.closest('.contact-option');
         const contactId = contactOption?.getAttribute('data-contact-id');
-        if (contactId) selectContact(contactId, container);
+        if (contactId) _selectContact(contactId, container);
     }, options);
 }
 
 /**
- * Initialize dropdown functionality
+ * Initializes dropdown functionality for a given container.
+ * Loads contacts, populates both dropdowns, sets up event delegation and outside-click handling.
  * @param {HTMLElement} container - The container element to scope queries
  * @param {HTMLElement} pageContainer - The main page container for global event delegation
  * @returns {void}
@@ -440,6 +259,7 @@ export function initializeDropdowns(container, pageContainer) {
     const signal = renewAbortController(container);
 
     activeContainer = container;
+    setActiveContainerRef(container);
 
     loadContacts().then(() => populateContactsDropdown(container));
     populateCategoriesDropdown(container);
@@ -448,12 +268,12 @@ export function initializeDropdowns(container, pageContainer) {
     pageContainer.addEventListener('click', (event) => closeDropdownOnClickOutside(event, container), {signal});
 
     container.addEventListener('keydown', (event) => {
-        if (event.key === 'Escape') closeAllDropdowns(null, container)
+        if (event.key === 'Escape') closeAllDropdowns(null, container);
     }, {signal});
 }
 
 /**
- * Renews the AbortController for a container, aborting any existing controller
+ * Renews the AbortController for a container, aborting any existing controller.
  * @param {HTMLElement} container - The container element to associate with the AbortController
  * @returns {AbortSignal} The signal from the new AbortController
  */
@@ -467,109 +287,17 @@ function renewAbortController(container) {
 }
 
 /**
- * Get selected contacts
- * @returns {Array} - Array of selected contacts
- */
-export function getSelectedContacts() {
-    return selectedContacts;
-}
-
-/**
- * Get selected category
- * @returns {object|null} - Selected category or null
- */
-export function getSelectedCategory() {
-    return selectedCategory;
-}
-
-/**
- * Clear selected contacts
- * @returns {void}
- */
-export function clearSelectedContacts() {
-    selectedContacts = [];
-}
-
-/**
- * Clear selected category
- * @returns {void}
- */
-export function clearSelectedCategory() {
-    selectedCategory = null;
-}
-
-/**
- * Reset dropdown state completely (contacts and category)
- * Use this when opening a new dialog/form to start fresh
+ * Resets dropdown state completely (both contacts and category).
+ * Use this when opening a new dialog/form to start fresh.
  * @returns {void}
  */
 export function resetDropdownState() {
-    selectedContacts = [];
-    selectedCategory = null;
+    _setSelectedContacts([]);
+    _setSelectedCategory(null);
 }
 
-/**
- * Waits for contacts to be loaded, then preselects the given member IDs
- * @param {string[]} memberIds - Array of contact IDs to preselect
- * @param {HTMLElement} container - The container element to scope queries
- * @returns {Promise<void>}
- */
-export async function preselectContacts(memberIds, container) {
-    if (!memberIds || memberIds.length === 0) return;
-    await loadContacts();
-    selectedContacts = [];
+// --- Re-exports from contactDropdownManager ---
+export {selectContact, preselectContacts, getSelectedContacts, setSelectedContacts} from "./contactDropdownManager.js";
 
-    container.querySelectorAll('.contact-option').forEach(option => {
-        option.classList.remove('selected');
-        const checkbox = option.querySelector('input[type="checkbox"]');
-        if (checkbox) checkbox.checked = false;
-    });
-
-    memberIds.forEach(contactId => {
-        handleMemberPreselection(contactId, container);
-    });
-    updateSelectedContactsDisplay(container);
-}
-
-/**
- * Handles the preselection of a single member by ID
- * @param {string} contactId - The contact ID to preselect
- * @param {HTMLElement} container - The container element to scope queries
- * @returns {void}
- */
-function handleMemberPreselection(contactId, container) {
-    const contact = contacts.find(c => c.id === contactId);
-    if (!contact) return;
-
-    selectedContacts.push(contact);
-
-    const contactOption = container.querySelector(`[data-contact-id="${contactId}"]`);
-    if (!contactOption) return;
-    contactOption.classList.add('selected');
-
-    const checkbox = contactOption.querySelector('input[type="checkbox"]');
-    if (checkbox) checkbox.checked = true;
-}
-
-/**
- * Preselects a category based on task type name
- * Sets the category without opening/toggling the dropdown
- * @param {string} taskType - The task type name (e.g., "User Story", "Technical Task")
- * @param {HTMLElement|Document} container - The container element to scope queries
- */
-export function preselectCategory(taskType, container) {
-    if (!taskType) return;
-    const categoryMap = {'User Story': 'user-story', 'Technical Task': 'technical'};
-    const categoryId = categoryMap[taskType];
-    if (!categoryId) return;
-    const category = categories.find(c => c.id === categoryId);
-    if (!category) return;
-    selectedCategory = category;
-    const categoryDisplay = container.querySelector('.category-display');
-    if (categoryDisplay) categoryDisplay.textContent = category.name;
-    container.querySelectorAll('.category-option').forEach(option => {
-        option.classList.remove('selected');
-    });
-    const selectedOption = container.querySelector(`[data-category-id="${categoryId}"]`);
-    if (selectedOption) selectedOption.classList.add('selected');
-}
+// --- Re-exports from categoryDropdownManager ---
+export {selectCategory, preselectCategory, getSelectedCategory, setSelectedCategory} from "./categoryDropdownManager.js";
